+++ OneBlink.SDK/models/AttachmentData.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace OneBlink.SDK.Model
 {
     public class AttachmentData : S3UploadResponse
@@ -18,7 +20,12 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        [Obsolete("Use tenant origin with the `path` property to construct an absolute URL")]
         public string url
+        {
+            get; set;
+        }
+        public string path
         {
             get; set;
         }